requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
Django==6.0.3
django-environ==0.13.0
django-ninja==1.6.0
django-tenants==3.10.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
redis==7.3.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
whitenoise==6.12.0
